+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.1
 
 * Introduce `Test.Tasty.Bench.Crossover`.
+* Introduce `guessPolynomialComplexity`.
 
 ## 0.1